Completed Add Increments: A-MoreOOP
